[refactor] fixes
